requirements.txt
yt-dlp==2023.12.30
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
beautifulsoup4==4.12.2
feedparser==6.0.10
yt-dlp==2023.12.30
